3.0.8 releases

+++ linpg/core/function.py
@@ -36,24 +36,28 @@ def loadDynamicImage(
 
 # 加载GIF格式图片
 def loadGif(
-    img_list_or_path: Union[list, tuple, str],
+    img_list_or_path: Union[str, tuple, list],
     position: tuple,
     size: tuple,
     updateGap: int = 1
-    ) -> GifObject:
-    if isinstance(img_list_or_path, str):
-        imgList:list = []
+    ) -> GifSurface:
+    imgList:list = []
+    #如果是gif文件
+    if isinstance(img_list_or_path, str) and img_list_or_path.endswith(".gif"):
         gif_image:object = Image.open(img_list_or_path)
         theFilePath:str = os.path.dirname(img_list_or_path)
         for i in range(gif_image.n_frames):
             gif_image.seek(i)
             pathTmp = os.path.join(theFilePath, "gifTempFileForLoading_{}.png".format(i))
             gif_image.save(pathTmp)
-            imgList.append(loadImg(pathTmp))
+            imgList.append(StaticImageSurface(pathTmp,0,0,size[0], size[1]))
             os.remove(pathTmp)
-        return GifObject(numpy.asarray(imgList), position[0], position[1], size[0], size[1], updateGap)
+    #如果是一个列表的文件路径
+    elif isinstance(img_list_or_path, (tuple, list)):
+        for image_path in img_list_or_path: imgList.append(StaticImageSurface(image_path,0,0,size[0], size[1]))
     else:
-        return GifObject(numpy.asarray(img_list_or_path), position[0], position[1], size[0], size[1], updateGap)
+        throwException("error", 'Invalid input for "img_list_or_path": {}'.format(img_list_or_path))
+    return GifSurface(numpy.asarray(imgList), position[0], position[1], size[0], size[1], updateGap)
 
 # 加载按钮
 def loadButton(
@@ -140,9 +144,32 @@ def isHover(
     else:
         return isHoverPygameObject(imgObject, objectPos, local_x, local_y)
 
-# 转换pygame的rect类
-def convert_rect(rect: pygame.Rect) -> Shape: return Shape(rect.x, rect.y, rect.width, rect.height)
+# 转换pygame的rect类至linpg引擎的shape类
+def convert_to_shape(rect: Union[Shape, pygame.Rect]) -> Shape:
+    #确认是pygame.Rect类再转换
+    if isinstance(rect, pygame.Rect):
+        return Shape(rect.x, rect.y, rect.width, rect.height)
+    #如果是Shape类，则没必要转换
+    elif isinstance(rect, Shape):
+        return rect
+    else:
+        throwException("error", 'The rect has to be "pygame.Rect" or at least "linpg.Shape", not "{}".'.format(type(rect)))
+
+# 转换linpg引擎的shape类至pygame的rect类
+def convert_to_rect(shape: Union[Shape, pygame.Rect]) -> pygame.Rect:
+    #确认是pygame.Rect类再转换
+    if isinstance(shape, pygame.Rect):
+        return shape
+    #如果是Shape类，则没必要转换
+    elif isinstance(shape, Shape):
+        return pygame.Rect(shape.left, shape.top, shape.width, shape.height)
+    else:
+        throwException("error", 'The shape has to be "linpg.Shape" or at least "pygame.Rect", not "{}".'.format(type(shape)))
 
 # 是否形状一样
 def is_same_shape(rect1: Union[Shape, pygame.Rect], rect2: Union[Shape, pygame.Rect]) -> bool:
     return rect1.x == rect2.x and rect1.y == rect2.y and rect1.width == rect2.width and rect1.height == rect2.height
+
+# 画正方形（的方块）
+def draw_rect(surface:pygame.Surface, color:any, rect:Union[pygame.Rect, Shape], width:int=0) -> None:
+    pygame.draw.rect(surface, color, convert_to_rect(rect), width)
